got vendor list working, began process of splitting into modules
