+++ src/templates/vendor-account.html
@@ -11,7 +11,6 @@
 <!-- I've copy pasted these lines of codes from profile.html and deleted that page -->
 {% extends 'layout.html' %}
 {% block head %}
-  <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='css/vendor-account.css') }}">
 {% endblock %}
 {% block body %}
   <div class="modal">
@@ -231,33 +230,7 @@
   </section>
 {% endblock %}
 {% block scripts %}
-<script>
-  var bookingViewer = new BookingViewer(
-    document.getElementById('bookingViewer'), {
-      dataFormat: "yyyy-mm-dd",
-      closeOnSelect: false
-    }
-  );
-</script>
 <% for (var chunk in htmlWebpackPlugin.files.chunks) { %>
-<script type="text/javascript" src="{{ url_for('static', filename='<%= htmlWebpackPlugin.files.chunks[chunk].entry %>') }}'"></script>
+<script type="text/javascript" src="{{ url_for('static', filename='<%= htmlWebpackPlugin.files.chunks[chunk].entry %>') }}"></script>
 <% } %>
-<script defer>
-  var loc, tag;
-  update_tag();
-  $(window).on('popstate', e => {
-    update_tag();
-  })
-  function update_tag() {
-    loc = window.location.href.split('#');
-    tag = loc.length > 1 ? loc[1] : '';
-    if (tag === "bookings") {
-      retrieveDates();
-    } else if (tag === "account") {
-      showAccountInfo();
-    } else if (!tag) {
-      retrieveDates();
-    }
-  }
-</script>
 {% endblock %}
