+++ src/vendor-plugins/framework.js
@@ -1,7 +1,6 @@
 import $ from "jquery";
 import progressively from "progressively";
 import "stickyfill";
-import { getVendorByType } from "../js/main.js";
 
 window.$ = window.jQuery = $;
 window.progressively = progressively;
