+++ templates/layout.html
@@ -151,11 +151,11 @@
       </div>
     </footer>
     
-      <script type="text/javascript" src="{{ url_for('static', filename='framework.94be2cb7f83e7498912f.js') }}"></script>
+      <script type="text/javascript" src="{{ url_for('static', filename='framework.3644d6847e8eb5e1b5c4.js') }}"></script>
     
-      <script type="text/javascript" src="{{ url_for('static', filename='main.94be2cb7f83e7498912f.js') }}"></script>
+      <script type="text/javascript" src="{{ url_for('static', filename='main.3644d6847e8eb5e1b5c4.js') }}"></script>
     
-      <script type="text/javascript" src="{{ url_for('static', filename='commonstyles.94be2cb7f83e7498912f.js') }}"></script>
+      <script type="text/javascript" src="{{ url_for('static', filename='commonstyles.3644d6847e8eb5e1b5c4.js') }}"></script>
     
     {% block scripts %}{% endblock %}
   </body>
